refactor with Lombok (#7)
Refactor the entity class with Lombok

+++ src/v/systems/type/DataType.java
@@ -15,7 +15,7 @@ public enum DataType {
     }
 
     public boolean equals(Integer id) {
-        return id != null && id.equals(typeId);
+        return id != null && id.equals((int)typeId);
     }
 
     public final byte getTypeId() {
